Create Graph and Details Page

+++ app/src/main/java/com/ahmed/ntg/data/remote/dto/LatestDto.kt
@@ -5,5 +5,5 @@ import java.util.*
 data class LatestDto(
     val base : String,
     val date: String,
-    val rates : Map<String, Double>
+    val rates : Map<String, Float>
 )
